+++ packages/liveblocks-core/src/comments/index.ts
@@ -20,18 +20,18 @@ type PartialNullable<T> = {
   [P in keyof T]?: T[P] | null | undefined;
 };
 
-export type CommentsApi<ThreadMetadata extends BaseMetadata> = {
-  getThreads(): Promise<ThreadData<ThreadMetadata>[]>;
+export type CommentsApi<TThreadMetadata extends BaseMetadata> = {
+  getThreads(): Promise<ThreadData<TThreadMetadata>[]>;
   createThread(options: {
     threadId: string;
     commentId: string;
-    metadata: ThreadMetadata | undefined;
+    metadata: TThreadMetadata | undefined;
     body: CommentBody;
-  }): Promise<ThreadData<ThreadMetadata>>;
+  }): Promise<ThreadData<TThreadMetadata>>;
   editThreadMetadata(options: {
-    metadata: PartialNullable<ThreadMetadata>;
+    metadata: PartialNullable<TThreadMetadata>;
     threadId: string;
-  }): Promise<ThreadData<ThreadMetadata>>;
+  }): Promise<ThreadData<TThreadMetadata>>;
   createComment(options: {
     threadId: string;
     commentId: string;
@@ -58,11 +58,11 @@ export type CommentsApi<ThreadMetadata extends BaseMetadata> = {
   }): Promise<CommentData>;
 };
 
-export function createCommentsApi<ThreadMetadata extends BaseMetadata>(
+export function createCommentsApi<TThreadMetadata extends BaseMetadata>(
   roomId: string,
   getAuthValue: () => Promise<AuthValue>,
   { serverEndpoint }: Options
-): CommentsApi<ThreadMetadata> {
+): CommentsApi<TThreadMetadata> {
   async function fetchJson<T>(
     endpoint: string,
     options?: RequestInit
@@ -116,12 +116,12 @@ export function createCommentsApi<ThreadMetadata extends BaseMetadata>(
     });
   }
 
-  async function getThreads(): Promise<ThreadData<ThreadMetadata>[]> {
+  async function getThreads(): Promise<ThreadData<TThreadMetadata>[]> {
     const response = await fetchApi(roomId, "/threads");
 
     if (response.ok) {
       const json = await (response.json() as Promise<{
-        data: ThreadData<ThreadMetadata>[];
+        data: ThreadData<TThreadMetadata>[];
       }>);
       return json.data;
     } else if (response.status === 404) {
@@ -140,10 +140,10 @@ export function createCommentsApi<ThreadMetadata extends BaseMetadata>(
     roomId: string;
     threadId: string;
     commentId: string;
-    metadata: ThreadMetadata | undefined;
+    metadata: TThreadMetadata | undefined;
     body: CommentBody;
   }) {
-    return fetchJson<ThreadData<ThreadMetadata>>("/threads", {
+    return fetchJson<ThreadData<TThreadMetadata>>("/threads", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
@@ -164,10 +164,10 @@ export function createCommentsApi<ThreadMetadata extends BaseMetadata>(
     threadId,
   }: {
     roomId: string;
-    metadata: PartialNullable<ThreadMetadata>;
+    metadata: PartialNullable<TThreadMetadata>;
     threadId: string;
   }) {
-    return fetchJson<ThreadData<ThreadMetadata>>(
+    return fetchJson<ThreadData<TThreadMetadata>>(
       `/threads/${encodeURIComponent(threadId)}/metadata`,
       {
         method: "POST",