+++ packages/liveblocks-react-comments/src/components/Comment.tsx
@@ -84,12 +84,12 @@ export interface CommentProps extends ComponentPropsWithoutRef<"div"> {
   /**
    * The event handler called when the comment is edited.
    */
-  onEdit?: (comment: CommentData) => void;
+  onCommentEdit?: (comment: CommentData) => void;
 
   /**
    * The event handler called when the comment is deleted.
    */
-  onDelete?: (comment: CommentData) => void;
+  onCommentDelete?: (comment: CommentData) => void;
 
   /**
    * The event handler called when clicking on the author.
@@ -275,8 +275,8 @@ export const Comment = forwardRef<HTMLDivElement, CommentProps>(
       showReactions = true,
       onAuthorClick,
       onMentionClick,
-      onEdit,
-      onDelete,
+      onCommentEdit,
+      onCommentDelete,
       overrides,
       additionalActions,
       additionalActionsClassName,
@@ -321,7 +321,7 @@ export const Comment = forwardRef<HTMLDivElement, CommentProps>(
     const handleEditSubmit = useCallback(
       ({ body }: ComposerSubmitComment, event: FormEvent<HTMLFormElement>) => {
         // TODO: Add a way to preventDefault from within this callback, to override the default behavior (e.g. showing a confirmation dialog)
-        onEdit?.(comment);
+        onCommentEdit?.(comment);
 
         event.preventDefault();
         setEditing(false);
@@ -331,18 +331,18 @@ export const Comment = forwardRef<HTMLDivElement, CommentProps>(
           body,
         });
       },
-      [comment, editComment, onEdit]
+      [comment, editComment, onCommentEdit]
     );
 
     const handleDelete = useCallback(() => {
       // TODO: Add a way to preventDefault from within this callback, to override the default behavior (e.g. showing a confirmation dialog)
-      onDelete?.(comment);
+      onCommentDelete?.(comment);
 
       deleteComment({
         commentId: comment.id,
         threadId: comment.threadId,
       });
-    }, [comment, deleteComment, onDelete]);
+    }, [comment, deleteComment, onCommentDelete]);
 
     const handleAuthorClick = useCallback(
       (event: MouseEvent<HTMLElement>) => {